(REF) loss: separate score functions from scalarisations
